+++ tasks/cleanup.js
@@ -1,27 +1,26 @@
 const cron = require("node-cron");
 const User = require("../models/userModel");
 
-
 const deleteOldNotifications = async () => {
-    const oneWeekAgo = new Date();
-    oneWeekAgo.setDate(oneWeekAgo.getDate() - 7); // Subtract 7 days
+  const twoDaysAgo = new Date();
+  twoDaysAgo.setDate(twoDaysAgo.getDate() - 2); // Subtract 1 day
 
-    try {
-        await User.updateMany(
-            {},
-            {
-                $pull: {
-                    notifications: {
-                        timestamp: { $lt: oneWeekAgo }
-                    }
-                }
-            }
-        );
+  try {
+    await User.updateMany(
+      {},
+      {
+        $pull: {
+          notifications: {
+            timestamp: { $lt: twoDaysAgo },
+          },
+        },
+      }
+    );
 
-        console.log("Old notifications deleted.");
-    } catch (error) {
-        console.error("Error deleting old notifications:", error);
-    }
+    console.log("Old notifications deleted.");
+  } catch (error) {
+    console.error("Error deleting old notifications:", error);
+  }
 };
 
 module.exports = deleteOldNotifications;
